[FIX] 커뮤니티 뷰모델 수정 중입니다.
