+++ app/src/main/AndroidManifest.xml
@@ -37,6 +37,10 @@
             </intent-filter>
         </activity>
 
+        <activity
+            android:name="com.google.android.gms.auth.api.signin.internal.SignInHubActivity"
+            android:screenOrientation="fullSensor"
+            android:windowSoftInputMode="stateAlwaysHidden|adjustPan" />
 
         <activity
             android:name=".ui.map.MapUserDetailActivity"
